pyproject.toml
[project]
name = "everland"
version = "0.1.0"
description = "환상의나라 - 슬랙 기반 멀티 게임 봇"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "jurigged>=0.6.1",
    "python-dotenv>=1.2.2",
    "slack-bolt>=1.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.11",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "E402"]
